Don't attempt to copy non-existing vardirs

+++ combinations.pl
@@ -353,18 +353,18 @@ sub doCombination {
             my $to = $workdir.'/vardir'.$s.'_'.$trial_id;
             say("[$thread_id] Copying $from to $to") if $logToStd;
             if (osWindows()) {
-                system("xcopy \"$from\" \"$to\" /E /I /Q");
+                system("xcopy \"$from\" \"$to\" /E /I /Q") if -e $from;
                 system("xcopy \"$from"."_slave\" \"$to\" /E /I /Q") if -e $from.'_slave';
                 open(OUT, ">$to/command");
                 print OUT $command;
                 close(OUT);
             } elsif ($command =~ m{--mem}) {
-                system("cp -r /dev/shm/var $to");
+                system("cp -r /dev/shm/var $to") if -e '/dev/shm/var';
                 open(OUT, ">$to/command");
                 print OUT $command;
                 close(OUT);
             } else {
-                system("cp -r $from $to");
+                system("cp -r $from $to") if -e $from;
                 system("cp -r $from"."_slave $to") if -e $from.'_slave';
 		if ($command =~ m/_epoch/) {
 			system("mv $epochcreadir $to");
